+++ VisualPlus/GDI.cs
@@ -188,6 +188,52 @@
             }
         }
 
+        /// <summary>Calculates a 5 point star.</summary>
+        /// <param name="originF"> The originF is the middle of the star.</param>
+        /// <param name="outerRadius">Radius of the surrounding circle.</param>
+        /// <param name="innerRadius">Radius of the circle for the "inner" points</param>
+        /// <returns>10 PointF array.</returns>
+        public static PointF[] Calculate5PointStar(PointF originF, float outerRadius, float innerRadius)
+        {
+            // Define some variables to avoid as much calculations as possible
+            // conversions to radians
+            const double Ang36 = Math.PI / 5.0; // 36Â° x PI/180
+            const double Ang72 = 2.0 * Ang36; // 72Â° x PI/180
+
+            // some sine and cosine values we need
+            var sin36 = (float)Math.Sin(Ang36);
+            var sin72 = (float)Math.Sin(Ang72);
+            var cos36 = (float)Math.Cos(Ang36);
+            var cos72 = (float)Math.Cos(Ang72);
+
+            // Fill array with 10 originF points
+            PointF[] pointsArray = { originF, originF, originF, originF, originF, originF, originF, originF, originF, originF };
+            pointsArray[0].Y -= outerRadius; // top off the star, or on a clock this is 12:00 or 0:00 hours
+            pointsArray[1].X += innerRadius * sin36;
+            pointsArray[1].Y -= innerRadius * cos36; // 0:06 hours
+            pointsArray[2].X += outerRadius * sin72;
+            pointsArray[2].Y -= outerRadius * cos72; // 0:12 hours
+            pointsArray[3].X += innerRadius * sin72;
+            pointsArray[3].Y += innerRadius * cos72; // 0:18
+            pointsArray[4].X += outerRadius * sin36;
+            pointsArray[4].Y += outerRadius * cos36; // 0:24 
+
+            // Phew! Glad I got that trig working.
+            pointsArray[5].Y += innerRadius;
+
+            // I use the symmetry of the star figure here
+            pointsArray[6].X += pointsArray[6].X - pointsArray[4].X;
+            pointsArray[6].Y = pointsArray[4].Y; // mirror point
+            pointsArray[7].X += pointsArray[7].X - pointsArray[3].X;
+            pointsArray[7].Y = pointsArray[3].Y; // mirror point
+            pointsArray[8].X += pointsArray[8].X - pointsArray[2].X;
+            pointsArray[8].Y = pointsArray[2].Y; // mirror point
+            pointsArray[9].X += pointsArray[9].X - pointsArray[1].X;
+            pointsArray[9].Y = pointsArray[1].Y; // mirror point
+
+            return pointsArray;
+        }
+
         /// <summary>Draws the control.</summary>
         /// <param name="control">The control to draw.</param>
         /// <param name="point">The point.</param>
