+++ VisualPlus/Styles/Visual.cs
@@ -39,7 +39,7 @@
             }
         }
 
-        public Gradient Background
+        public Gradient BackProgress
         {
             get
             {
